# packages/contracts/openapi.yaml
openapi: 3.1.0

# Контракт HTTP API. Описаны только реализованные маршруты: описание того, чего
# нет, читается как обещание и приводит к клиенту, написанному под
# несуществующий сервер. Проверка contracts сверяет этот файл с маршрутами
# приложения в обе стороны, поэтому новые endpoints добавляются сюда вместе с
# кодом, а не после него.

info:
  title: Персональная ИИ-Система развития — API
  version: 0.1.0
  description: |
    Состояние: health, sessions/dev-login, me и команды целей/заданий.
    Telegram initData auth, webhook, bootstrap/pull ещё не реализованы.
    Synthetic login разрешён только в разработке. OpenAPI отражает текущие
    маршруты; целевая архитектура Telegram описана в docs/06 и docs/14.
    Пробелы command payload validation/versions перечислены в docs/15.

    Доступ закрыт по умолчанию: маршрут, не перечисленный как публичный,
    требует access-токена.

servers:
  - url: http://127.0.0.1:3000
    description: Локальная разработка

paths:
  /health:
    get:
      summary: Проба живости
      description: |
        Отвечает, пока процесс жив. Намеренно не обращается к базе данных:
        иначе перезапуск приложения зависел бы от доступности базы, и
        перезапускали бы не то, что сломалось.
      operationId: getHealth
      responses:
        '200':
          description: Процесс отвечает
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [status, environment, uptimeSeconds]
                properties:
                  status:
                    const: ok
                  environment:
                    type: string
                    enum: [development, test, production]
                  uptimeSeconds:
                    type: integer
                    minimum: 0

  /auth/dev-login:
    post:
      summary: Вход по синтетической личности (только разработка)
      description: |
        Существует лишь при включённом `DEV_AUTH_ENABLED` вне production. При
        выключенном флаге отвечает 404: отключённый маршрут не должен
        подтверждать своё существование. Проверка Apple identity token не
        реализована и остаётся незакрытой частью P1-02.
      operationId: devLogin
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                subject:
                  type: string
                  description: Идентификатор синтетической личности; из него детерминированно выводится пользователь.
      responses:
        '201':
          description: Сессия выпущена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssuedSession'
        '404':
          description: Вход для разработки выключен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/telegram:
    post:
      summary: Вход через Telegram Mini App
      description: |
        Клиент передаёт исходную строку `initData`, полученную от Telegram.
        Токен бота клиенту не передаётся и в теле запроса не принимается.

        Подпись — единственное доказательство личности: `initDataUnsafe`,
        username и отображаемое имя её не подтверждают. Проверяется HMAC по
        официальной спецификации, возраст подписи (не старше пяти минут),
        расхождение часов вперёд (не больше тридцати секунд) и принадлежность
        подписанного пользователя к списку допущенных. Пустой список означает
        «никого».

        Строка — краткоживящее предъявительское доказательство: подпись не
        делает украденную строку безопасной. Одно доказательство обменивается
        ровно на одну семью сессий; повторное предъявление отклоняется с
        `initdata_already_used`. Если ответ потерян, клиент открывает Mini App
        заново и получает свежее доказательство — ошибка не подменяется
        успехом.

        Без настроенного токена бота маршрут отвечает 404: ненастроенный вход не
        подтверждает своё существование.
      operationId: loginWithTelegram
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [init_data]
              properties:
                init_data:
                  type: string
                  maxLength: 8192
                installation_id:
                  type: string
                  format: uuid
      responses:
        '201':
          description: Выдана сессия
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/IssuedSession'
                  - type: object
                    properties:
                      user_id:
                        type: string
                        format: uuid
        '400':
          description: |
            Запрос собран неправильно: строка длиннее допустимого, не
            разбирается, содержит дубли ключей, не содержит подписи или
            подписанного пользователя.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: |
            Личность не подтверждена: подпись не совпала (`initdata_forged`),
            доказательство просрочено (`initdata_expired`) или из будущего
            (`initdata_from_future`), уже обменено (`initdata_already_used`),
            человек не допущен к пилоту (`user_not_allowed`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Аккаунт удалён или приостановлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Вход через Telegram не настроен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Сбой инфраструктуры; вход можно повторить новым запуском
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/refresh:
    post:
      summary: Обновление сессии
      description: |
        Предъявляется refresh-токен, а не access. Повторное предъявление уже
        использованного токена означает, что он есть у двоих: семья сессий
        отзывается целиком и возвращается `token_reuse_detected`.

        401 означает именно приговор сессии: токен неизвестен, истёк или
        отозван. Сбой инфраструктуры отвечает 503 и не должен приниматься за
        недействительную сессию — клиент по 401 стирает локальный журнал, и
        временная недоступность базы стоила бы несинхронизированных записей.
      operationId: refreshSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
      responses:
        '200':
          description: Выданы новые токены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssuedSession'
        '400':
          description: Токен не передан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Токен неизвестен, истёк, отозван или использован повторно
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: |
            Сбой инфраструктуры. Сессия не отменена: повторить позже, локальный
            журнал не стирать.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/logout:
    post:
      summary: Отзыв семьи сессий
      operationId: logout
      security:
        - accessToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [family_id]
              properties:
                family_id:
                  type: string
                  format: uuid
      responses:
        '204':
          description: Семья отозвана
        '400':
          description: Не передан family_id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Нет действующего access-токена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /me:
    get:
      summary: Текущий пользователь
      operationId: getMe
      security:
        - accessToken: []
      responses:
        '200':
          description: Пользователь, подтверждённый access-токеном
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [user_id]
                properties:
                  user_id:
                    type: string
                    format: uuid
        '401':
          description: Нет действующего access-токена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /commands:
    post:
      summary: Выполнить команду
      description: |
        Единая точка изменений. Пользователь берётся из access-токена: в
        конверте нет поля пользователя, и схема закрыта для посторонних полей,
        поэтому подставить чужой идентификатор нечем.

        Повторная доставка того же `command_id` не создаёт второго эффекта и
        возвращает прежнюю квитанцию с `duplicate: true`. Тот же `command_id` с
        другим смыслом — конфликт: подтверждать выполнение того, чего сервер не
        делал, нельзя. Смысл команды — это её вид, цель, ожидаемая версия,
        заявленная зависимость и нагрузка; идентификатор устройства и время
        клиента при повторе меняются законно и в сравнение не входят.

        Цель (`aggregate_id`) и ожидаемая версия (`expected_version`) — поля
        конверта. Их дубликаты в нагрузке приняты только при совпадении:
        расхождение означает, что клиент и сервер понимают команду по-разному,
        и выбрать один вариант молча нельзя.

        Изменяющая команда обязана назвать оба поля: без версии «отметить
        выполненным» означает «выполнено, что бы там сейчас ни было», и команда
        по устаревшему экрану применится к состоянию, которого человек не
        видел. Создающая команда присылает оба поля пустыми — предыдущей версии
        у неё нет, а собственный идентификатор объекта, если клиент его
        выбирает, лежит в нагрузке.

        Тело проверяется по `schemas/command-envelope.schema.json` — тому же
        файлу, по которому клиент собирает запрос. Конверт описывает только
        транспорт, поэтому нагрузка дополнительно проверяется по закрытой схеме
        своего вида из `schemas/commands/`: неизвестное поле отклоняется, а не
        теряется молча.

        `depends_on_command_id` пока не поддержан и отклоняется явно
        (`unsupported_dependency`): выполнить команду, для которой клиент
        заявил предшественника, значит нарушить порядок, который он считает
        гарантированным.

        Завершение и частичное выполнение записывают факт действия: фактические
        объём или длительность, отрезок времени и источник. Мера должна
        совпадать с мерой спецификации задания. Если объём не прислан, факт
        записывается как неизмеренный — запланированное значение не
        подставляется, иначе отметку не отличить от измерения. Частичное
        выполнение без объёма отклоняется: его не отличить от невыполнения.
        `correct_activity` уточняет уже записанный факт, заменяя его под тем же
        корнем, а не создавая второе действие.
      operationId: executeCommand
      security:
        - accessToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './schemas/command-envelope.schema.json'
      responses:
        '200':
          description: Команда выполнена либо распознана как повтор
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [command_id, committed_seq, duplicate, result]
                properties:
                  command_id:
                    type: string
                    format: uuid
                  committed_seq:
                    type: string
                    description: Номер изменения; передаётся строкой, чтобы не терять точность больших чисел.
                  duplicate:
                    type: boolean
                  result:
                    type: object
        '400':
          description: |
            Конверт не прошёл проверку (`bad_request`), неизвестный вид команды
            (`unknown_command_kind`), нагрузка не прошла закрытую схему или
            доменную проверку (`invalid_payload`), цель конверта расходится с
            целью нагрузки (`target_mismatch`), нарушена политика версии
            (`version_required`), заявлена неподдержанная зависимость
            (`unsupported_dependency`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Нет действующего access-токена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: |
            Тот же command_id уже выполнен с другим смыслом (`command_id_reused`),
            объект изменился с момента чтения (`version_conflict`), переход
            недопустим в текущем состоянии (`invalid_transition`) или такая
            строка уже существует (`already_exists`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /health/ready:
    get:
      summary: Проба готовности
      description: |
        Проверяет доступность базы данных. При недоступной базе отвечает 503 и
        не раскрывает причину: строка подключения и параметры сервера не
        предназначены для внешнего клиента.
      operationId: getHealthReady
      responses:
        '200':
          description: База отвечает, запросы можно направлять
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [status, database]
                properties:
                  status:
                    const: ready
                  database:
                    const: up
        '503':
          description: База недоступна
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [status, database]
                properties:
                  status:
                    const: not_ready
                  database:
                    const: down

components:
  securitySchemes:
    accessToken:
      type: http
      scheme: bearer
      description: |
        Непрозрачный access-токен сессии, живёт 15 минут. Проверяется по базе,
        поэтому отзыв действует немедленно.

  schemas:
    IssuedSession:
      type: object
      additionalProperties: true
      required: [access_token, access_expires_at, refresh_token, refresh_expires_at, family_id]
      properties:
        access_token:
          type: string
        access_expires_at:
          type: string
          format: date-time
        refresh_token:
          type: string
        refresh_expires_at:
          type: string
          format: date-time
        family_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid

    Error:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: string
        detail:
          type: string
          description: |
            Пояснение для клиента там, где ошибка в самом запросе. Внутренности
            сервера сюда не попадают: живая проверка однажды показала в ответе
            имя колонки базы.
        request_id:
          type: string
          description: |
            Идентификатор запроса при ошибке сервера. Возвращается вместо текста
            ошибки: человек может сослаться на свой случай, а внутренности
            остаются в логах.